removing using Slack to determine if user is a teamlead
This Slack API method has been depreciated. There is a different way of making it work, but it seems that simply having selectable boxes for is a teamlead or not is more intuitive and simple anyway. This should only change twice a year or so and would not need to be updated that often, and makes the interface more clear that those properties are actually editable.